sync: fuse previous high-risk plugin, artifact, and wallet work
Integrate password transport encryption, int64 wallet quota, task-plugin
sandbox, signed artifacts, video proxy, and plugin marketplace while
preserving local image-task, payment, failover, and billing behavior.

+++ controller/token.go
@@ -310,7 +310,7 @@ func AddToken(c *gin.Context) {
 			return
 		}
 		maxQuotaValue := common.QuotaFromFloat(1000000000 * common.QuotaPerUnit)
-		if token.RemainQuota > maxQuotaValue {
+		if token.RemainQuota > int64(maxQuotaValue) {
 			common.ApiErrorI18n(c, i18n.MsgTokenQuotaExceedMax, map[string]any{"Max": maxQuotaValue})
 			return
 		}
@@ -408,7 +408,7 @@ func UpdateToken(c *gin.Context) {
 			return
 		}
 		maxQuotaValue := common.QuotaFromFloat(1000000000 * common.QuotaPerUnit)
-		if token.RemainQuota > maxQuotaValue {
+		if token.RemainQuota > int64(maxQuotaValue) {
 			common.ApiErrorI18n(c, i18n.MsgTokenQuotaExceedMax, map[string]any{"Max": maxQuotaValue})
 			return
 		}
