+++ model/user_cache.go
@@ -18,7 +18,7 @@ type UserBase struct {
 	Id          int    `json:"id"`
 	Group       string `json:"group"`
 	Email       string `json:"email"`
-	Quota       int    `json:"quota"`
+	Quota       int64  `json:"quota"`
 	Status      int    `json:"status"`
 	Role        int    `json:"role"`
 	Username    string `json:"username"`
@@ -209,7 +209,7 @@ func cacheDecrUserQuota(userId int, delta int64) error {
 // syncCreditUserQuotaCache 在授信事务（充值/兑换等）提交后同步把增量补进缓存
 // 余额。预扣以缓存值为准（存在期间），授信不能绕过它，否则新到账的额度在
 // 缓存过期前不可用；缓存未命中无需处理，下次读取会从已提交的数据库余额水合。
-func syncCreditUserQuotaCache(userId int, quota int, operation string) {
+func syncCreditUserQuotaCache(userId int, quota int64, operation string) {
 	if quota <= 0 {
 		return
 	}
@@ -227,7 +227,7 @@ func getUserGroupCache(userId int) (string, error) {
 	return cache.Group, nil
 }
 
-func getUserQuotaCache(userId int) (int, error) {
+func getUserQuotaCache(userId int) (int64, error) {
 	cache, err := GetUserCache(userId)
 	if err != nil {
 		return 0, err
@@ -260,7 +260,7 @@ func updateUserStatusCache(userId int, status bool) error {
 	return updateUserCacheField(userId, "Status", statusInt)
 }
 
-func updateUserQuotaCache(userId int, quota int) error {
+func updateUserQuotaCache(userId int, quota int64) error {
 	if !common.RedisEnabled {
 		return nil
 	}
